"Inheritance Practice - Employees, Managers, Trainees"

+++ src/com/teksystems/Trainees.java
@@ -1,7 +1,7 @@
 package com.teksystems;
 
 public class Trainees extends Employee{
-    public Trainees(long employeeId, String employeeName, String employeeAddress, long employeePhone, long basicSalary) {
+    public Trainees(long employeeId, String employeeName, String employeeAddress, long employeePhone, double basicSalary) {
         super(employeeId, employeeName, employeeAddress, employeePhone, basicSalary);
     }
 }
